pyproject.toml
[project]
name = "cloud-diagram-cli"
version = "0.3.0"
description = "Modular CLI tool for generating multi-cloud infrastructure diagrams using DOT/Graphviz format"
authors = [
    {name = "Cloud Diagram CLI", email = "noreply@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
keywords = ["aws", "azure", "gcp", "cloud", "infrastructure", "diagram", "graphviz", "cli", "multi-cloud"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: Software Development :: Documentation",
]
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.35.0",
    "diagrams>=0.23.0",
    "graphviz>=0.20.0",
    "ruamel.yaml>=0.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
]

[project.scripts]
cloud-diagram = "cloud_diagram.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cloud_diagram"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]